Automatic Loading of yml-Tasks from Install Folder
The added functionality checks for the existence of the AUTO_PROJ_ROOT
environment variable and subsequently the existence of the path to the
yml model files. If either of those not exists, the user will get get
notfified via a messagebox stating the problem.

Furthermore, the complete code file has been formatted with the clang
formatter, which has worked amazingly well in combination with
KDevelop... in your face vim and emacs :)

Signed-off-by: Bernd Langpap <Bernd.Langpap@dfki.de>