Added support for transaction notification
This has been added to baseGraph to allow transaction framing. E.g. when
loading a file, the entire event can be started with a TRANSACTION_START,
and ended with a TRANSACTION_STOP event. In this way, activities that
belong together can be bundled in the observers, and allow to operate on
the final structure.